+++ src/webR/robj-main.ts
@@ -24,6 +24,7 @@ export type RDouble = RProxy<RWorker.RDouble>;
 export type RComplex = RProxy<RWorker.RComplex>;
 export type RCharacter = RProxy<RWorker.RCharacter>;
 export type RList = RProxy<RWorker.RList>;
+export type RDataFrame = RProxy<RWorker.RDataFrame>;
 export type RRaw = RProxy<RWorker.RRaw>;
 export type RCall = RProxy<RWorker.RCall>;
 // RFunction proxies are callable